// src/servers/ws/dtos/message.dto.ts
import { z } from 'zod';
import messageType from '@/servers/ws/enums/message.type';

export interface MessageChatTypeInterface {
  type: messageType.Chat;
  to: string;
  text: string;
}

export interface MessageAuthenticateTypeInterface {
  type: messageType.Authenticate;
  token: string;
}

export const WsMessage = z.discriminatedUnion('type', [
  z.object({
    type: z.literal(messageType.Chat),
    to: z.string(),
    text: z.string(),
  }),
  z.object({
    type: z.literal(messageType.Authenticate),
    token: z.string(),
  }),
  z.object({
    type: z.literal(messageType.Subscribe),
    topic: z.string(),
  }),
  z.object({
    type: z.literal(messageType.Unsubscribe),
    topic: z.string(),
  }),
]);